+++ internal/controller/user_controller.go
@@ -165,8 +165,8 @@ func getCurrentUserIDFromContext(c *gin.Context) int {
 }
 
 type OrderResponseModel struct {
-	Number     string `json:"number"`
-	Status     string `json:"status"`
-	Accrual    int    `json:"accrual,omitempty"`
-	UploadedAt string `json:"uploaded_at" `
+	Number     string  `json:"number"`
+	Status     string  `json:"status"`
+	Accrual    float64 `json:"accrual,omitempty"`
+	UploadedAt string  `json:"uploaded_at" `
 }
